refactor(payments): Move getCustomerPayPalSubscriptions to SubscriptionManager
